+++ www/Library/content_widgets/Comments/Comments.php
@@ -6,12 +6,24 @@
  */
 namespace Library\content_widgets\Comments;
 
-use Library\views\AutoWidgetList\AutoWidgetList;
+use Library\views\AutoWidgetList\AutoWidgetList,
+    Boolive\values\Rule;
 
 class Comments extends AutoWidgetList
 {
+    public function getInputRule()
+    {
+        return Rule::arrays(array(
+            'show_title' => Rule::bool()->default(true)->required(),
+            'REQUEST' => Rule::arrays(array(
+                'object' => Rule::entity()->default($this->object)->required(),
+            ))
+        ));
+    }
+
     public function work($v = array())
     {
-        echo 'Comments';
+        $v['show_title'] = $this->_input['show_title'];
+        return parent::work($v);
     }
 }